requirements2.txt
beautifulsoup4==4.9.0
blis==0.4.1
boto==2.49.0
boto3==1.13.16
botocore==1.16.16
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.1
cymem==2.0.3
Cython==0.29.14
danlp==0.0.7
docutils==0.15.2
future==0.18.2
gensim==3.8.3
gitdb==4.0.4
GitPython==3.1.1
googletrans==2.4.0
ibm-db==3.0.1
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
joblib==0.15.1
langdetect==1.0.8
mock==4.0.1
murmurhash==1.0.2
nltk==3.5
normalize==2.0.2
nose2==0.9.2
pandas==1.0.3
plac==1.1.3
preshed==3.0.2
public==2019.4.13
py-translate==1.0.3
pycodestyle==2.6.0
pyconll==2.2.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
query-string==2019.4.13
regex==2020.5.14
requests==2.23.0
richenum==1.3.1
s3transfer==0.3.3
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
six==1.14.0
smart-open==2.0.0
smmap==3.0.2
soupsieve==2.0
spacy==2.3.0
srsly==1.0.2
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.0.0
tqdm==4.46.0
unique==0.1
urllib3==1.25.9
wasabi==0.6.0
zipp==3.1.0